style: restructure files to ease and support separating single and compound variants
