--- packages/core/src/convert/toDashed.js
@@ -1,5 +1,0 @@
-/** Returns a filled value with a dash prefix. */
-export const toLeadDashed = value => value ? '-' + value : '' // prettier-ignore
-
-/** Returns a filled value with a dash prefix. */
-export const toTailDashed = value => value ? value + '-' : '' // prettier-ignore
