+++ packages/core/src/convert/toHash.js
@@ -1,14 +1,23 @@
-/** Returns a unique hash for an value. */
-export const toHash = (
-	/** @type {string} Hashable value. */
-	value,
-) => {
-	/** @see https://stackoverflow.com/questions/7616461/generate-a-hash-from-string-in-javascript/52171480?#comment115384813_52171480 */
-	for (var index = value.length, hash = 9; index; ) {
-		hash = Math.imul(hash ^ value.charCodeAt(--index), 9 ** 9)
-	}
+const toAlphabeticChar = (/** @type {number} */ code) => String.fromCharCode(code + (code > 25 ? 39 : 97))
 
-	const hashString = (hash ^ (hash >>> 9)).toString(36).slice(-5)
+const toAlphabeticName = (/** @type {number} */ code) => {
+	let name = ''
+	let x
 
-	return hashString
+	for (x = Math.abs(code); x > 52; x = (x / 52) | 0) name = toAlphabeticChar(x % 52) + name
+
+	return toAlphabeticChar(x % 52) + name
 }
+
+const toPhash = (/** @type {number} */ h, /** @type {string} */ x) => {
+	let i = x.length
+	while (i) h = (h * 33) ^ x.charCodeAt(--i)
+	return h
+}
+
+export const toHash = (/** @type {string} */ value) => toAlphabeticName(
+	toPhash(
+		5381,
+		JSON.stringify(value)
+	) >>> 0
+) // prettier-ignore
