+++ packages/core/src/convert/toTailDashed.js
@@ -1,0 +1,2 @@
+/** Returns a filled value with a dash prefix. */
+export const toTailDashed = (/** @type {string} */ value) => value ? value + '-' : '' // prettier-ignore
