--- packages/core/src/createObject.js
@@ -1,2 +1,0 @@
-/** @type {<T>(target: T, source: any) => T} */
-export const createObject = (target, source) => Object.defineProperties(target, Object.getOwnPropertyDescriptors(source))
