--- packages/core/src/stringify/sizeProps.js
@@ -1,15 +1,0 @@
-/** CSS Properties whose value include a sizing keyword. */
-export default {
-	blockSize: 1,
-	height: 1,
-	inlineSize: 1,
-	maxBlockSize: 1,
-	maxHeight: 1,
-	maxInlineSize: 1,
-	maxWidth: 1,
-	minBlockSize: 1,
-	minHeight: 1,
-	minInlineSize: 1,
-	minWidth: 1,
-	width: 1,
-}
