+++ packages/core/src/utility/define.js
@@ -1,0 +1,2 @@
+/** @type {<T>(target: T, source: any) => T} */
+export const define = (target, source) => Object.defineProperties(target, Object.getOwnPropertyDescriptors(source))
