+++ packages/core/tests/basic.js
@@ -119,7 +119,7 @@ describe('Basic', () => {
 			`--stitches{--:2 c-dataoT}@media{` +
 				`.c-dataoT{color:DodgerBlue}` +
 			`}` +
-			`--stitches{--:4 c-dataoT-icaIZdx-css}@media{` +
+			`--stitches{--:5 c-dataoT-icaIZdx-css}@media{` +
 				`.c-dataoT-icaIZdx-css{color:Crimson}` +
 			`}`
 		)
@@ -146,7 +146,7 @@ describe('Basic', () => {
 
 		expect(component1of1).toBeInstanceOf(Function)
 		expect(className1of1).toBe('c-imTdEZ c-imTdEZ-caIZdx-variant')
-		expect(cssString1of1).toBe(`--stitches{--:2 c-imTdEZ}@media{.c-imTdEZ{font-size:100%}}--stitches{--:3 c-imTdEZ-caIZdx-variant}@media{.c-imTdEZ-caIZdx-variant{color:Crimson}}`)
+		expect(cssString1of1).toBe(`--stitches{--:2 c-imTdEZ}@media{.c-imTdEZ{font-size:100%}}--stitches{--:4 c-imTdEZ-caIZdx-variant}@media{.c-imTdEZ-caIZdx-variant{color:Crimson}}`)
 	})
 
 	test('Functionality of css() — utils', () => {
