+++ packages/core/tests/component-variants.js
@@ -58,14 +58,14 @@ describe('Variants', () => {
 		const expression1CssText = '.c-PJLV-Gaggi-variant{font-size:16px}'
 
 		expect(expression1.className).toBe('c-PJLV c-PJLV-Gaggi-variant')
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-Gaggi-variant}@media{${expression1CssText}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-Gaggi-variant}@media{${expression1CssText}}`)
 
 		const expression2 = component({ color: 'blue' })
 
 		const expression2CssText = '.c-PJLV-kaCQqN-variant{background-color:dodgerblue;color:white}'
 
 		expect(expression2.className).toBe('c-PJLV c-PJLV-kaCQqN-variant')
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-Gaggi-variant c-PJLV-kaCQqN-variant}@media{${expression1CssText + expression2CssText}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-Gaggi-variant c-PJLV-kaCQqN-variant}@media{${expression1CssText + expression2CssText}}`)
 	})
 
 	test('Renders a component with 2 matching variants', () => {
@@ -97,7 +97,7 @@ describe('Variants', () => {
 		const expressionSizeSmallCssText = '.c-PJLV-Gaggi-variant{font-size:16px}'
 		const expressionLevel1CssText = '.c-PJLV-iRwLiB-variant{padding:0.5em}'
 
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-Gaggi-variant c-PJLV-iRwLiB-variant}@media{${expressionSizeSmallCssText + expressionLevel1CssText}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-Gaggi-variant c-PJLV-iRwLiB-variant}@media{${expressionSizeSmallCssText + expressionLevel1CssText}}`)
 	})
 
 	test('Renders a component with a 2 matching variants and 1 matching compound', () => {
@@ -110,7 +110,7 @@ describe('Variants', () => {
 		const expressionCompoundCssText = '.c-PJLV-cChFtv-variant{transform:scale(1.2)}'
 
 		expect(expression.className).toBe(`c-PJLV c-PJLV-kaCQqN-variant c-PJLV-Gaggi-variant c-PJLV-cChFtv-variant`)
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-kaCQqN-variant c-PJLV-Gaggi-variant c-PJLV-cChFtv-variant}@media{${expressionColorBlueCssText + expressionSizeSmallCssText + expressionCompoundCssText}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-kaCQqN-variant c-PJLV-Gaggi-variant c-PJLV-cChFtv-variant}@media{${expressionColorBlueCssText + expressionSizeSmallCssText + expressionCompoundCssText}}`)
 	})
 }) // prettier-ignore
 
@@ -164,7 +164,7 @@ describe('Variants with defaults', () => {
 		const expression = component()
 
 		expect(expression.className).toBe('c-PJLV c-PJLV-Gaggi-variant')
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-Gaggi-variant}@media{.c-PJLV-Gaggi-variant{font-size:16px}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-Gaggi-variant}@media{.c-PJLV-Gaggi-variant{font-size:16px}}`)
 	})
 
 	test('Renders a component with the default variant explicitly applied', () => {
@@ -173,7 +173,7 @@ describe('Variants with defaults', () => {
 		const expression = component({ size: 'small' })
 
 		expect(expression.className).toBe('c-PJLV c-PJLV-Gaggi-variant')
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-Gaggi-variant}@media{.c-PJLV-Gaggi-variant{font-size:16px}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-Gaggi-variant}@media{.c-PJLV-Gaggi-variant{font-size:16px}}`)
 	})
 
 	test('Renders a component with the non-default variant explicitly applied', () => {
@@ -182,7 +182,7 @@ describe('Variants with defaults', () => {
 		const expression = component({ size: 'large' })
 
 		expect(expression.className).toBe('c-PJLV c-PJLV-hsYHIj-variant')
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-hsYHIj-variant}@media{.c-PJLV-hsYHIj-variant{font-size:24px}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-hsYHIj-variant}@media{.c-PJLV-hsYHIj-variant{font-size:24px}}`)
 	})
 
 	test('Renders a component with the default variant applied and a different variant explicitly applied', () => {
@@ -192,7 +192,7 @@ describe('Variants with defaults', () => {
 
 		expect(expression.className).toBe('c-PJLV c-PJLV-Gaggi-variant c-PJLV-iRwLiB-variant')
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-Gaggi-variant c-PJLV-iRwLiB-variant}@media{` +
+			`--stitches{--:4 c-PJLV-Gaggi-variant c-PJLV-iRwLiB-variant}@media{` +
 				// implicit size:small
 				`.c-PJLV-Gaggi-variant{font-size:16px}` +
 				// explicit level:1
@@ -208,7 +208,7 @@ describe('Variants with defaults', () => {
 
 		expect(expression.className).toBe('c-PJLV c-PJLV-kaCQqN-variant c-PJLV-Gaggi-variant c-PJLV-cChFtv-variant')
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-kaCQqN-variant c-PJLV-Gaggi-variant c-PJLV-cChFtv-variant}@media{` +
+			`--stitches{--:4 c-PJLV-kaCQqN-variant c-PJLV-Gaggi-variant c-PJLV-cChFtv-variant}@media{` +
 				// explicit color:blue
 				`.c-PJLV-kaCQqN-variant{background-color:dodgerblue;color:white}` +
 				// implicit size:small
@@ -225,7 +225,7 @@ describe('Variants with defaults', () => {
 		const className = `${component}`
 
 		expect(className).toBe('c-PJLV')
-		expect(getCssString()).toBe('--stitches{--:3 c-PJLV-Gaggi-variant}@media{.c-PJLV-Gaggi-variant{font-size:16px}}')
+		expect(getCssString()).toBe('--stitches{--:4 c-PJLV-Gaggi-variant}@media{.c-PJLV-Gaggi-variant{font-size:16px}}')
 	})
 }) // prettier-ignore
 
@@ -295,7 +295,7 @@ describe('Conditional variants', () => {
 		const componentSmallCssText = `.${componentSmallClassName}{font-size:16px}`
 
 		expect(component({ size: 'small' }).className).toBe([componentClassName, componentSmallClassName].join(' '))
-		expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-Gaggi-variant}@media{${componentSmallCssText}}`)
+		expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-Gaggi-variant}@media{${componentSmallCssText}}`)
 	})
 
 	test('Renders a component with one conditional variant on one breakpoint applied', () => {
@@ -304,7 +304,7 @@ describe('Conditional variants', () => {
 
 		expect(component({ size: { '@bp1': 'small' } }).className).toBe(`c-PJLV c-PJLV-iVKIeV-variant`)
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-iVKIeV-variant}@media{` +
+			`--stitches{--:4 c-PJLV-iVKIeV-variant}@media{` +
 				`@media (max-width: 767px){.c-PJLV-iVKIeV-variant{font-size:16px}}` +
 			`}`
 		)
@@ -321,7 +321,7 @@ describe('Conditional variants', () => {
 
 		expect(component({ size: { '@bp1': 'small', '@bp2': 'large' } }).className).toBe([componentClassName, componentSmallBp1ClassName, componentLargeBp2ClassName].join(' '))
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
+			`--stitches{--:4 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
 				componentSmallBp1CssText +
 				componentLargeBp2CssText +
 			`}`
@@ -334,7 +334,7 @@ describe('Conditional variants', () => {
 
 		expect(component({ size: { '@bp1': 'small', '@bp2': 'large' } }).className).toBe(`c-PJLV c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant`)
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
+			`--stitches{--:4 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
 				`@media (max-width: 767px){.c-PJLV-iVKIeV-variant{font-size:16px}}` +
 				`@media (min-width: 768px){.c-PJLV-bUkcYv-variant{font-size:24px}}` +
 			`}`
@@ -342,7 +342,7 @@ describe('Conditional variants', () => {
 
 		expect(component({ size: { '@bp1': 'small', '@bp2': 'large' } }).className).toBe(`c-PJLV c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant`)
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
+			`--stitches{--:4 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
 				`@media (max-width: 767px){.c-PJLV-iVKIeV-variant{font-size:16px}}` +
 				`@media (min-width: 768px){.c-PJLV-bUkcYv-variant{font-size:24px}}` +
 			`}`
@@ -350,7 +350,7 @@ describe('Conditional variants', () => {
 
 		expect(component({ size: { '@bp1': 'small', '@bp2': 'large' } }).className).toBe(`c-PJLV c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant`)
 		expect(getCssString()).toBe(
-			`--stitches{--:3 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
+			`--stitches{--:4 c-PJLV-iVKIeV-variant c-PJLV-bUkcYv-variant}@media{` +
 				`@media (max-width: 767px){.c-PJLV-iVKIeV-variant{font-size:16px}}` +
 				`@media (min-width: 768px){.c-PJLV-bUkcYv-variant{font-size:24px}}` +
 			`}`
@@ -383,7 +383,7 @@ describe('Conditional variants', () => {
 			).toBe('c-PJLV c-PJLV-gjWYHE-variant c-PJLV-fzmUzy-variant')
 
 			expect(getCssString()).toBe(
-				`--stitches{--:3 c-PJLV-gjWYHE-variant c-PJLV-fzmUzy-variant}@media{` +
+				`--stitches{--:4 c-PJLV-gjWYHE-variant c-PJLV-fzmUzy-variant}@media{` +
 					`@media (max-width:767.9375px){.c-PJLV-gjWYHE-variant{font-size:16px}}` +
 					`@media (min-width:768px){.c-PJLV-fzmUzy-variant{font-size:24px}}` +
 				`}`
@@ -415,7 +415,7 @@ describe('Conditional variants', () => {
 			).toBe('c-PJLV c-PJLV-gjWYHE-variant c-PJLV-fzmUzy-variant')
 
 			expect(getCssString()).toBe(
-				`--stitches{--:3 c-PJLV-gjWYHE-variant c-PJLV-fzmUzy-variant}@media{` +
+				`--stitches{--:4 c-PJLV-gjWYHE-variant c-PJLV-fzmUzy-variant}@media{` +
 					`@media (max-width:767.9375px){.c-PJLV-gjWYHE-variant{font-size:16px}}` +
 					`@media (min-width:768px){.c-PJLV-fzmUzy-variant{font-size:24px}}` +
 				`}`
@@ -460,7 +460,7 @@ describe('Variant pairing types', () => {
 			`--stitches{--:2 c-foEXqW}@media{` +
 				`.c-foEXqW{--component:true}` +
 			`}` +
-			`--stitches{--:3 c-foEXqW-iloXEi-variant}@media{` +
+			`--stitches{--:4 c-foEXqW-iloXEi-variant}@media{` +
 				`.c-foEXqW-iloXEi-variant{--test-boolean:true}` +
 			`}`
 		)
@@ -476,7 +476,7 @@ describe('Variant pairing types', () => {
 			`--stitches{--:2 c-foEXqW}@media{` +
 				`.c-foEXqW{--component:true}` +
 			`}` +
-			`--stitches{--:3 c-foEXqW-iloXEi-variant}@media{` +
+			`--stitches{--:4 c-foEXqW-iloXEi-variant}@media{` +
 				`.c-foEXqW-iloXEi-variant{--test-boolean:true}` +
 			`}`
 		)
@@ -492,7 +492,7 @@ describe('Variant pairing types', () => {
 			`--stitches{--:2 c-foEXqW}@media{` +
 				`.c-foEXqW{--component:true}` +
 			`}` +
-			`--stitches{--:3 c-foEXqW-brOaTK-variant}@media{` +
+			`--stitches{--:4 c-foEXqW-brOaTK-variant}@media{` +
 				`@media (min-width: 640px){` +
 					`.c-foEXqW-brOaTK-variant{--test-boolean:true}` +
 				`}` +
@@ -510,7 +510,7 @@ describe('Variant pairing types', () => {
 			`--stitches{--:2 c-foEXqW}@media{` +
 				`.c-foEXqW{--component:true}` +
 			`}` +
-			`--stitches{--:3 c-foEXqW-brOaTK-variant}@media{` +
+			`--stitches{--:4 c-foEXqW-brOaTK-variant}@media{` +
 				`@media (min-width: 640px){` +
 					`.c-foEXqW-brOaTK-variant{--test-boolean:true}` +
 				`}` +
