+++ packages/core/tests/issue-450.js
@@ -56,28 +56,28 @@ describe('Issue #450', () => {
 			const { component1, getCssString } = getFreshComponents()
 			const render = component1()
 			expect(render.className).toBe(`c-PJLV c-PJLV-gmqXFB-variant`)
-			expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-gmqXFB-variant}@media{.c-PJLV-gmqXFB-variant{color:red}}`)
+			expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-gmqXFB-variant}@media{.c-PJLV-gmqXFB-variant{color:red}}`)
 		})
 
 		test('Render component1({ color: "blue" }) as blue, assigned from props', () => {
 			const { component1, getCssString } = getFreshComponents()
 			const render = component1({ color: 'blue' })
 			expect(render.className).toBe(`c-PJLV c-PJLV-kydkiA-variant`)
-			expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-kydkiA-variant}@media{.c-PJLV-kydkiA-variant{color:blue}}`)
+			expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-kydkiA-variant}@media{.c-PJLV-kydkiA-variant{color:blue}}`)
 		})
 
 		test('Render component1({ color: "red" }) as red, assigned from props', () => {
 			const { component1, getCssString } = getFreshComponents()
 			const render = component1({ color: 'red' })
 			expect(render.className).toBe(`c-PJLV c-PJLV-gmqXFB-variant`)
-			expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-gmqXFB-variant}@media{.c-PJLV-gmqXFB-variant{color:red}}`)
+			expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-gmqXFB-variant}@media{.c-PJLV-gmqXFB-variant{color:red}}`)
 		})
 
 		test('Render component1({ color: { "@media (width >= 640px)": "blue" } }) as red then blue, inherited from defaultVariants, assigned from props', () => {
 			const { component1, getCssString } = getFreshComponents()
 			const render = component1({ color: { '@media (min-width: 640px)': 'blue' } })
 			expect(render.className).toBe(`c-PJLV c-PJLV-gmqXFB-variant c-PJLV-bBevdw-variant`)
-			expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-gmqXFB-variant c-PJLV-bBevdw-variant}@media{.c-PJLV-gmqXFB-variant{color:red}@media (min-width: 640px){.c-PJLV-bBevdw-variant{color:blue}}}`)
+			expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-gmqXFB-variant c-PJLV-bBevdw-variant}@media{.c-PJLV-gmqXFB-variant{color:red}@media (min-width: 640px){.c-PJLV-bBevdw-variant{color:blue}}}`)
 		})
 
 		test('Render component2() as orange, inherited from defaultVariants', () => {
@@ -85,7 +85,7 @@ describe('Issue #450', () => {
 			const render = component2()
 
 			expect(render.className).toBe(`c-PJLV c-PJLV-vMTTG-variant`)
-			expect(getCssString()).toBe(`--stitches{--:3 c-PJLV-vMTTG-variant}@media{.c-PJLV-vMTTG-variant{color:orange}}`)
+			expect(getCssString()).toBe(`--stitches{--:4 c-PJLV-vMTTG-variant}@media{.c-PJLV-vMTTG-variant{color:orange}}`)
 		})
 
 		test('Render component2({ color: { "@media (width >= 640px)": "blue" } }) as orange then blue, inherited from defaultVariants, assigned from props', () => {
@@ -93,7 +93,7 @@ describe('Issue #450', () => {
 			const render = component2({ color: { '@media (min-width: 640px)': 'blue' } })
 			expect(render.className).toBe(`c-PJLV c-PJLV-bBevdw-variant c-PJLV-vMTTG-variant`)
 			expect(getCssString()).toBe(
-				`--stitches{--:3 c-PJLV-bBevdw-variant c-PJLV-vMTTG-variant}@media{` +
+				`--stitches{--:4 c-PJLV-bBevdw-variant c-PJLV-vMTTG-variant}@media{` +
 					`@media (min-width: 640px){.c-PJLV-bBevdw-variant{color:blue}}` +
 					`.c-PJLV-vMTTG-variant{color:orange}` +
 				`}`
@@ -163,7 +163,7 @@ describe('Issue #450', () => {
 				`--stitches{--:2 c-jyxqjt}@media{` +
 					`.c-jyxqjt{--component:1}` +
 				`}` +
-				`--stitches{--:3 c-jyxqjt-ilDyRi-variant}@media{` +
+				`--stitches{--:4 c-jyxqjt-ilDyRi-variant}@media{` +
 					`.c-jyxqjt-ilDyRi-variant{--color:lightBlue}` +
 				`}`
 			)
@@ -177,7 +177,7 @@ describe('Issue #450', () => {
 				`--stitches{--:2 c-jyxqjt}@media{` +
 					`.c-jyxqjt{--component:1}` +
 				`}` +
-				`--stitches{--:3 c-jyxqjt-cOChOn-variant}@media{` +
+				`--stitches{--:4 c-jyxqjt-cOChOn-variant}@media{` +
 					`.c-jyxqjt-cOChOn-variant{--appearance:secondary}` +
 				`}`
 			)
@@ -191,7 +191,7 @@ describe('Issue #450', () => {
 				`--stitches{--:2 c-jyxqjt}@media{` +
 					`.c-jyxqjt{--component:1}` +
 				`}` +
-				`--stitches{--:3 c-jyxqjt-cOChOn-variant c-jyxqjt-ilDyRi-variant c-jyxqjt-gYqlvA-variant}@media{` +
+				`--stitches{--:4 c-jyxqjt-cOChOn-variant c-jyxqjt-ilDyRi-variant c-jyxqjt-gYqlvA-variant}@media{` +
 					`.c-jyxqjt-cOChOn-variant{--appearance:secondary}` +
 					`.c-jyxqjt-ilDyRi-variant{--color:lightBlue}` +
 					`.c-jyxqjt-gYqlvA-variant{--compound:appearance secondary / color lightBlue}` +
@@ -208,7 +208,7 @@ describe('Issue #450', () => {
 					`.c-jyxqjt{--component:1}` +
 					`.c-dkRcuu{--component:2}` +
 				`}` +
-				`--stitches{--:3 c-jyxqjt-cOChOn-variant c-jyxqjt-ilDyRi-variant c-jyxqjt-gYqlvA-variant}@media{` +
+				`--stitches{--:4 c-jyxqjt-cOChOn-variant c-jyxqjt-ilDyRi-variant c-jyxqjt-gYqlvA-variant}@media{` +
 					`.c-jyxqjt-cOChOn-variant{--appearance:secondary}` +
 					`.c-jyxqjt-ilDyRi-variant{--color:lightBlue}` +
 					`.c-jyxqjt-gYqlvA-variant{--compound:appearance secondary / color lightBlue}` +
@@ -225,7 +225,7 @@ describe('Issue #450', () => {
 					`.c-jyxqjt{--component:1}` +
 					`.c-dkRcuu{--component:2}` +
 				`}` +
-				`--stitches{--:3 c-jyxqjt-cOChOn-variant c-jyxqjt-ilDyRi-variant c-jyxqjt-gYqlvA-variant}@media{` +
+				`--stitches{--:4 c-jyxqjt-cOChOn-variant c-jyxqjt-ilDyRi-variant c-jyxqjt-gYqlvA-variant}@media{` +
 					`.c-jyxqjt-cOChOn-variant{--appearance:secondary}` +
 					`.c-jyxqjt-ilDyRi-variant{--color:lightBlue}` +
 					`.c-jyxqjt-gYqlvA-variant{--compound:appearance secondary / color lightBlue}` +
